+++ ADEPO/server/mount_coord_spots.h
@@ -2,13 +2,13 @@
 #define MOUNT_COORD_SPOTS_H
 
 #include <iostream>
-#include "point3f.h"
+#include "point3d.h"
 #include "bcam.h"
 
 class MountCoordSpots
 {
 public:
-    MountCoordSpots(BCAM bcam, Point3f coord1, Point3f coord2) :
+    MountCoordSpots(BCAM bcam, Point3d coord1, Point3d coord2) :
         mBCAM(bcam), mCoord1(coord1), mCoord2(coord2) {};
     virtual ~MountCoordSpots() {};
 
@@ -16,8 +16,8 @@ public:
     Prism getPrism() const { return mBCAM.getPrism(); }
     QString getName() const { return getBCAM().getName()+"_"+getPrism().getName(); }
 
-    Point3f getCoord1() const {return mCoord1; }
-    Point3f getCoord2() const {return mCoord2; }
+    Point3d getCoord1() const {return mCoord1; }
+    Point3d getCoord2() const {return mCoord2; }
 
     //methodes
     void print() {
@@ -33,8 +33,8 @@ public:
 protected:
 private:
     BCAM mBCAM;
-    Point3f mCoord1;
-    Point3f mCoord2;
+    Point3d mCoord1;
+    Point3d mCoord2;
 };
 
 #endif // MOUNT_COORD_SPOTS_H
